Move `MatchOnVecItems` into `Matches` lint pass
