+++ clippy_lints/src/lib.register_lints.rs
@@ -258,7 +258,6 @@ store.register_lints(&[
     map_err_ignore::MAP_ERR_IGNORE,
     map_unit_fn::OPTION_MAP_UNIT_FN,
     map_unit_fn::RESULT_MAP_UNIT_FN,
-    match_on_vec_items::MATCH_ON_VEC_ITEMS,
     match_result_ok::MATCH_RESULT_OK,
     match_str_case_mismatch::MATCH_STR_CASE_MISMATCH,
     matches::COLLAPSIBLE_MATCH,
@@ -267,6 +266,7 @@ store.register_lints(&[
     matches::MATCH_AS_REF,
     matches::MATCH_BOOL,
     matches::MATCH_LIKE_MATCHES_MACRO,
+    matches::MATCH_ON_VEC_ITEMS,
     matches::MATCH_OVERLAPPING_ARM,
     matches::MATCH_REF_PATS,
     matches::MATCH_SAME_ARMS,
